+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/configuration/DynamicConfigFactory.java
@@ -8,20 +8,21 @@
 
 package org.opensearch.identity.configuration;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.greenrobot.eventbus.EventBus;
+import org.greenrobot.eventbus.EventBusBuilder;
+import org.opensearch.client.Client;
+import org.opensearch.common.settings.Settings;
+import org.opensearch.identity.User;
+import org.opensearch.identity.configuration.model.InternalUsersModel;
+import org.opensearch.identity.extensions.ExtensionSecurityConfigStore;
+import org.opensearch.identity.realm.InternalUsersStore;
+import org.opensearch.threadpool.ThreadPool;
+
 import java.nio.file.Path;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
-
-import org.opensearch.identity.User;
-import org.opensearch.identity.configuration.model.InternalUsersModel;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
-import org.opensearch.client.Client;
-import org.opensearch.common.settings.Settings;
-import org.opensearch.identity.realm.InternalUsersStore;
-import org.opensearch.threadpool.ThreadPool;
-import org.greenrobot.eventbus.EventBus;
-import org.greenrobot.eventbus.EventBusBuilder;
 
 public class DynamicConfigFactory implements ConfigurationChangeListener {
 
@@ -50,6 +51,7 @@ public class DynamicConfigFactory implements ConfigurationChangeListener {
         this.configPath = configPath;
 
         registerDCFListener(InternalUsersStore.getInstance());
+        registerDCFListener(ExtensionSecurityConfigStore.getInstance());
         this.cr.subscribeOnChange(this);
     }
 
@@ -72,12 +74,15 @@ public class DynamicConfigFactory implements ConfigurationChangeListener {
 
         }
 
-        final InternalUsersModel ium;
-        ium = new InternalUsersModelV1((SecurityDynamicConfiguration<User>) internalusers);
+        final InternalUsersModel ium = new InternalUsersModel((SecurityDynamicConfiguration<User>) internalusers);
 
         eventBus.post(ium);
+        eventBus.post(cr.getConfiguration(CType.EXTENSIONSECURITY));
 
-        initialized.set(true);
+        if (!isInitialized()) {
+            initialized.set(true);
+            cr.initializeExtensionsSecurity();
+        }
 
     }
 
@@ -91,37 +96,5 @@ public class DynamicConfigFactory implements ConfigurationChangeListener {
 
     public void unregisterDCFListener(Object listener) {
         eventBus.unregister(listener);
-    }
-
-    private static class InternalUsersModelV1 extends InternalUsersModel {
-
-        private final SecurityDynamicConfiguration<User> internalUserSecurityDynamicConfiguration;
-
-        public InternalUsersModelV1(SecurityDynamicConfiguration<User> internalUserSecurityDynamicConfiguration) {
-            super();
-            this.internalUserSecurityDynamicConfiguration = internalUserSecurityDynamicConfiguration;
-        }
-
-        @Override
-        public User getUser(String username) {
-            return internalUserSecurityDynamicConfiguration.getCEntry(username);
-        }
-
-        @Override
-        public boolean exists(String username) {
-            return internalUserSecurityDynamicConfiguration.exists(username);
-        }
-
-        @Override
-        public Map<String, String> getAttributes(String username) {
-            User tmp = internalUserSecurityDynamicConfiguration.getCEntry(username);
-            return tmp == null ? null : tmp.getAttributes();
-        }
-
-        @Override
-        public String getHash(String username) {
-            User tmp = internalUserSecurityDynamicConfiguration.getCEntry(username);
-            return tmp == null ? null : tmp.getHash();
-        }
     }
 }